Double distortion: fixed click issues
The transition between gains is linear, but the compensation gain
(if any) should be reciprocal. To simplify computations, it is
currently linearly ramping too. This cause problem when switching
immediately to a very low gain, because the gain and compensation gain
product suddenly goes very high in the middle of the course, causing a
distinct click.

This issues occurs mostly when switching presets, so handling
a special case after buffer clearing was enough to fix it.