Fixed a corruption bug with clang + misc details
In the clang atomic library, weak and strong CAS implementations
are different (at least for 32-bit ARM). The lock-free queue algorithm
relies on the exact success or failure of this primitive, so when the
weak CAS fails spuriously, it leads to data corruption. Replacing
weak CAS with strong CAS fixes the problem.

Also useless volatile keywords were removed, and ptrdiff_t replaced
with intptr_t when relevant.