Fixed compilation errors
Probably more to come later. Until C++17, static const(expr) member
variables are not systematically inlined and should also have a body
if the linker needs it. Everything depends on the compilation options.
Generally this is OK with maximum optimisations, but sometimes debug
builds require a body for some random constants.

To be safe, all constants should have a body, but this is a lot of work
and these constraints will vanish once C++17 will be adopted in the
source tree. So we'll fix compilation errors when they show up.