fstb::ToolsSimd: refactoring
The Vect*32 types are now separate classes in the V*32 form.
This allows adding constructors and conversion functions, which is an
important step to get unique template function definitions instead of
separated scalar/vector ones.
Moreover, the operator implementation is cleaner.

Not tested yet on ARM, fixes will probably follow.